WIP handling existing object storage files
handle string->uuid conversion better
existing object keys can be processed regardless of key format (uuid or string)
This means we will get duplicates because we have no keying mechanism between the two storage systems. I think the way around this is to only upload images that don't have a uuid as their key (so key left as filename or whatever), and assume that a uuid key has already been handled. Fixing this will require some sort of consistent identifier for files--maybe a hash of the metadata?