Improve handling of HEIC images
the version of tika that pantomime depends on is too old and say that .HEIC files are "video/quicktime" and not "image/heic". Update to version with fix.
Some files had their mime types as uppercase, so convert to lowercase before comparing
Start with pantomime.media/image? check, fall back to custom code if that doesn't work